Move resizedImage function to dedicated module ImageResize
